+++ arduino-ide-Support/[1.6.x]-hardware(library)/hardware/Microduino/avr/libraries/_09_DATA_Microduino_Protocol_SoftSer/Microduino_Protocol.cpp
@@ -20,7 +20,7 @@ byte getChecksum(byte length, byte cmd, byte mydata[]) {
   } */
 
 
-void Protocol::begin(uint16_t _baud) {
+void Protocol::begin(uint32_t _baud) {
   P_Serial->begin(_baud);
   delay(20);
 }
